refactor: moved factory to sentence module and renamed some files
