+++ src/index.spec.ts
@@ -1,5 +1,5 @@
 import defaults from './defaults';
-import { createSentence, configure } from '.';
+import { createSentence, configure } from './index';
 
 describe('sentence-engine', () => {
   /**
