+++ src/index.ts
@@ -1,7 +1,6 @@
 // eslint-disable-next-line no-unused-vars
 import { Template, Vocabulary, Options, Configuration, StringResolvable, WeightedEntry } from '../types';
-import { SentenceFactory } from './factory';
-import { Sentence } from './sentence';
+import { Sentence, SentenceFactory } from './sentence';
 import defaults from './defaults';
 
 const Facade = new SentenceFactory(defaults.templates, defaults.vocabulary, {});
