+++ src/sentence/SentenceFactory.spec.ts
