+++ src/sentence/index.ts
@@ -1,1 +1,2 @@
 export { Sentence } from './Sentence';
+export { SentenceFactory } from './SentenceFactory';
